prevent error-masking keyError
When doing a C-MOVE with DCM4CHE with a misconfigured IP address, a
KeyError results when trying to parse the DIMSE C-MOVE-Rsp. The KeyError
is because the message does not contain normal attributes, instead has a
Failure status, Error comment ETC.

This change means that it is possible to read the real error from the
Status object emitted by the SOPClass SCU -- missing attributes are set
to None.